close the channel if the funding tx never confirms
If we are fundee and after 2 weeks the funding tx is still not
confirmed, then we give up waiting and consider the channel closed.

If we are funder, we never give up until the funding tx is double spent,
and we periodically republish it.

This applies to states `WAIT_FOR_FUNDING_CONFIRMED` and `CLOSING`.

TODO: implement for Electrum

Fixes #1029.